+++ lib/src/core/dialog_support.dart
@@ -6,11 +6,12 @@ abstract interface class DialogSupport {
   ///
   /// Returns the index of the selected option, or null if the dialog was
   /// cancelled or not implemented.
-  Future<int?> showSingleSelectDialog(List<String> options);
+  Future<int?> showSingleSelectDialog(List<String> options, {String? title});
 
   /// Shows a multi select dialog with the given [options].
   ///
   /// Returns the indices of the selected options, or null if the dialog was
   /// cancelled or not implemented.
-  Future<Set<int>?> showMultiSelectDialog(List<String> options);
+  Future<Set<int>?> showMultiSelectDialog(List<String> options,
+      {String? title});
 }